+++ ChangeLog.md
@@ -1,3 +1,15 @@
+# v1.2.2
+* Localize the tooltips
+* Added a schema setting to help with any future data migrations
+
+# v1.2.1
+* Added tooltip for cover adjustments
+
+# v1.2.0
+* Look for built-in cover effect if not using PF2e Perception
+* If enabled, let PF2e Perception autocalculate cover at combat start
+* Show cover adjustments on initiative message
+
 # v1.1.1
 * Improvements to README.md and table format